move future utils
